Fix TestWithTempDir test to pass on Mac
Fix TestWithTmpeDir test that comapre os.getcwd() and tmpdir.
os.getcwd() returns the result of resolving the symbolic link. But tmpdir
absolute path. Testing fails on Mac because the /tmp directory is a symbolic
link. Comparing the results of resolving tempdir will pass the test. But
RustPython does not implement lstat, so it does not pass, but in CPython The
test passes.