Skip string char traversal on each next in python str iterator
Previosly, each call to `__next__` on a str iterator did a `chars().count()`
to find the length of the full string and traversed the chars of the string
until the expected char position from the start.

This approach ignores the char position and only cares about the byte position,
allowing it to only act on the specific part of the string binary it's interested
in.