Wrap `ItemFn` and #[pyattr] with `static_cell!`
When we have to add #[pyattr] with function, that function must be
called only once when module is imported.

In previous, we did it with add static_cell for each function or skip
it(undesired).

I thought these static_cell stmt for all of each function in modules are
quite duplicated.

Thus, I add wrapping code when the given item is `ItemFn` and have `#[pyattr]`.

Signed-off-by: snowapril <sinjihng@gmail.com>