Make stdio a feature
When embedding RustPython into a graphical windows application (where
windows_subsystem = "windows"), it fails to create the RustPython VM.

This happens because the VM will try to open stdin/stdout/stderr when
initializing, but stdio isn't available for /SUBSYSTEM:WINDOWS apps.

This commit makes stdio a feature, so it can be disabled when needed.

https://doc.rust-lang.org/reference/runtime.html#the-windows_subsystem-attribute
https://learn.microsoft.com/en-us/cpp/build/reference/subsystem-specify-subsystem?view=msvc-170
https://asawicki.info/news_1768_ways_to_print_and_capture_text_output_of_a_process